// src/modules/inpatient/nursing-care-plan/models/nursing-care-plan.model.ts
export interface IRoomModel {
  ID_Kamar: number;
  ID_KelasKamar: string;
  Nama_Kamar: string;
  Kode_Kamar: string;
  Tarif: number;
}
// ...
export class NursingCarePlan extends DataModel {
  kamar : Array<IRoomModel>;
  form?: NursingCarePlanModel;
  inform_consent: IProvisionOfInformationFormModel;

  constructor(req: INursingCarePlanModel) {
    super(req);
    this.kamar = req.kamar
    this.form = new NursingCarePlanModel(req.form);
    this.inform_consent = req.inform_consent;
  }
}